pyproject.toml
[project]
name = "microtonal-viz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "crepe>=0.0.15",
    "librosa>=0.10.1",
    "numpy==1.26.4",
    "pesto-pitch>=1.0.0",
    "polars==1.5.0",
    "pydub==0.25.1",
    "pygame-ce>=2.5.1",
    "pygame-gui>=0.6.12",
    "simpleaudio==1.0.4",
    "tensorflow>=2.15.0",
]
